Add GoogleMapsTileProvider class
