+++ geemap/common.py
@@ -16169,162 +16169,6 @@ def is_on_aws():
     return on_aws
 
 
-def google_map_tiles(
-    map_type: str = None,
-    language: str = "en-Us",
-    region: str = "US",
-    api_key: Optional[str] = None,
-    **kwargs: Any,
-):
-    """
-    Generates Google Map tiles using the provided parameters. To get an API key
-        and enable Map Tiles API, visit
-        https://developers.google.com/maps/get-started#create-project.
-        You can set the API key using the environment variable `MAPS_API_KEY`
-        or by passing it as an argument.
-
-    Args:
-        map_type (str, optional): The type of map to generate. Options are
-            'roadmap', 'satellite', 'terrain', 'hybrid', 'traffic', 'streetview'.
-            Defaults to None, which generates all map types, and returns a
-            dictionary of TileProvider objects.
-        language (str, optional): An IETF language tag that specifies the
-            language used to display information on the tiles, such as 'zh-Cn'.
-            Defaults to 'en-Us'.
-        region (str, optional): A Common Locale Data Repository region
-            identifier (two uppercase letters) that represents the physical
-            location of the user. Defaults to 'US'.
-        api_key (str, optional): The API key to use for the Google Maps API.
-            If not provided, it will try to get it from the environment or
-            Colab user data. Defaults to None.
-        **kwargs: Additional parameters to pass to the map generation. For more
-            info, visit https://bit.ly/3UhbZKU
-
-    Raises:
-        ValueError: If the API key is not provided and cannot be found in the
-            environment or Colab user data.
-        ValueError: If the map_type is not one of the allowed types.
-
-    Example:
-        >>> import geemap
-        >>> m = geemap.Map()
-        >>> basemap = geemap.google_map_tiles(map_type='roadmap',
-            language="en-Us", region="US", scale="scaleFactor2x", highDpi=True)
-        >>> m.add_basemap(basemap)
-
-    Returns:
-        TileProvider | dict: A TileProvider object with the generated map, or a
-            dictionary of TileProvider objects for all map types,
-            or None if the map could not be generated.
-    """
-
-    from xyzservices import TileProvider
-
-    if api_key is None:
-
-        if in_colab_shell():
-            from google.colab import userdata
-
-            api_key = userdata.get("MAPS_API_KEY")
-        else:
-            api_key = os.environ.get("MAPS_API_KEY")
-
-    if api_key is None:
-        raise ValueError(
-            "API key is required to access Google Maps API. To get an API key "
-            "and enable Map Tiles API, visit "
-            "https://developers.google.com/maps/get-started#create-project"
-        )
-
-    allowed_map_types = [
-        "roadmap",
-        "satellite",
-        "terrain",
-        "hybrid",
-        "traffic",
-        "streetview",
-    ]
-
-    # Support map type as a string with or without 'google.',
-    # such as 'Google Roadmap', 'Google.Roadmap', or 'Roadmap'
-    if isinstance(map_type, str):
-        map_type = map_type.lower().replace("google.", "").replace("google", "").strip()
-
-        if map_type not in allowed_map_types:
-            raise ValueError(
-                "mapType must be one of 'roadmap', 'satellite', 'terrain', "
-                "'hybrid', 'traffic', 'streetview'"
-            )
-
-    tile_args = {}
-
-    # Define the parameters for each map type
-    for m_type in allowed_map_types:
-
-        mapType = m_type
-        layerTypes = None
-
-        if m_type == "hybrid":
-            mapType = "satellite"
-            layerTypes = ["layerRoadmap"]
-        elif m_type == "terrain":
-            layerTypes = ["layerRoadmap"]
-        elif m_type == "traffic":
-            mapType = "roadmap"
-            layerTypes = ["layerTraffic"]
-        elif m_type == "streetview":
-            mapType = "roadmap"
-            layerTypes = ["layerStreetview"]
-
-        tile_args[m_type] = {
-            "mapType": mapType,
-            "language": language,
-            "region": region,
-            "layerTypes": layerTypes,
-            **kwargs,
-        }
-
-        if tile_args[m_type].get("layerTypes") is None:
-            del tile_args[m_type]["layerTypes"]
-
-    gmap_providers = {}
-
-    for m_type in allowed_map_types:
-
-        # If map_type is provided, only generate the specified map
-        if map_type is not None and m_type != map_type:
-            continue
-
-        args = tile_args[m_type]
-        response = requests.post(
-            f"https://tile.googleapis.com/v1/createSession?key={api_key}",
-            headers={"Content-Type": "application/json"},
-            json=args,
-        )
-
-        if response.status_code == 200:
-            res = response.json()
-            gmap_provider = TileProvider(
-                {
-                    "url": f"https://tile.googleapis.com/v1/2dtiles/{{z}}/{{x}}/{{y}}?session={res['session']}&key={{accessToken}}",
-                    "attribution": f"© Google {m_type.capitalize()}",
-                    "accessToken": api_key,
-                    "name": f"Google.{m_type.capitalize()}",
-                    "ext": res["imageFormat"],
-                    "tileSize": res["tileWidth"],
-                }
-            )
-            gmap_providers[m_type] = gmap_provider
-        else:
-            display(response.text)
-            gmap_provider = None
-
-    if map_type is None:  # Return all map types
-        return gmap_providers
-    else:  # Return the specified map type
-        return gmap_providers.get(map_type)
-
-
 def google_maps_api_key(lookup_key: str = "MAPS_API_KEY") -> Optional[str]:
     """
     Retrieves the Google Maps API key from the environment or Colab user data.
